requirements.txt
Django==5.2.7
djangorestframework==3.16.1
drf-yasg==1.21.10
django-cors-headers==4.9.0
django-environ==0.12.0
pillow==11.3.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
PyYAML==6.0.3
